backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
boto3
supabase
google-api-python-client
google-auth
apscheduler
